WINDUP-3720 Decoupled JSON extract generation from reports rendering (#1606)
* WINDUP-3720 Decoupled JSON extract generation from reports rendering

* Format with IntellJ

---------
